split misc into logic and interactive
