+++ src/interactive.c
@@ -1,0 +1,17 @@
+#include "interactive.h"
+
+#include <stdio.h>
+
+void greeting() {
+    puts("hello player!\n"
+         "welcome to a simple carcassonne based game!\n"
+         "for usage run: carcassonne help\n");
+}
+
+void usage() {
+    puts("usage: carcassonne [tiles-list-file] [board-file]\n"
+         "tiles-list-file and board-file should be flies in current directory\n"
+         "if both tiles-list-file and board-file specified run in auto mode\n"
+         "if only tiles-list given use list specified in interactive mode\n"
+         "if none file specified use default tile list for interactive mode\n");
+}
