+++ src/interactive.h
@@ -1,0 +1,15 @@
+#ifndef INTERACTIVE_H
+#define INTERACTIVE_H
+/** @file interactive.h */
+
+/**
+ * print usage.
+ */
+void usage();
+
+/**
+ * print greeting.
+ */
+void greeting();
+
+#endif
